+++ engine/test/util/fast_vector_tests.cpp
@@ -17,14 +17,14 @@ TEST(FastVector_Tests, Size_Trivial_AfterDefaultConstruction_Zero)
 {
   FastVector<double> vector;
 
-  EXPECT_EQ(vector.Size(), 0);
+  EXPECT_EQ(vector.Size(), 0u);
 }
 
 TEST(FastVector_Tests, Capacity_Trivial_AfterDefaultConstruction_Zero)
 {
   FastVector<double> vector;
 
-  EXPECT_EQ(vector.Capacity(), 0);
+  EXPECT_EQ(vector.Capacity(), 0u);
 }
 
 TEST(FastVector_Tests, PushBack_Trivial_Single_SizeIncrease)
@@ -33,7 +33,7 @@ TEST(FastVector_Tests, PushBack_Trivial_Single_SizeIncrease)
 
   vector.PushBack(0);
 
-  EXPECT_EQ(vector.Size(), 1);
+  EXPECT_EQ(vector.Size(), 1u);
 }
 
 TEST(FastVector_Tests, PushBack_NonTrivial_Single_SizeIncrease)
@@ -42,7 +42,7 @@ TEST(FastVector_Tests, PushBack_NonTrivial_Single_SizeIncrease)
 
   vector.PushBack("0");
 
-  EXPECT_EQ(vector.Size(), 1);
+  EXPECT_EQ(vector.Size(), 1u);
 }
 
 TEST(FastVector_Tests, PushBack_Trivial_Double_SizeIncrease)
@@ -52,7 +52,7 @@ TEST(FastVector_Tests, PushBack_Trivial_Double_SizeIncrease)
   vector.PushBack(0);
   vector.PushBack(0);
 
-  EXPECT_EQ(vector.Size(), 2);
+  EXPECT_EQ(vector.Size(), 2u);
 }
 
 TEST(FastVector_Tests, PushBack_NonTrivial_Double_SizeIncrease)
@@ -62,7 +62,7 @@ TEST(FastVector_Tests, PushBack_NonTrivial_Double_SizeIncrease)
   vector.PushBack("0");
   vector.PushBack("0");
 
-  EXPECT_EQ(vector.Size(), 2);
+  EXPECT_EQ(vector.Size(), 2u);
 }
 
 TEST(FastVector_Tests, PushBack_Trivial_Many_SizeIncrease)
@@ -178,7 +178,7 @@ TEST(FastVector_Tests, PopBack_Trivial_Single_SizeDecrease)
   vector.PushBack(0);
   vector.PopBack();
 
-  EXPECT_EQ(vector.Size(), 0);
+  EXPECT_EQ(vector.Size(), 0u);
   EXPECT_TRUE(vector.Empty());
 }
 
@@ -191,7 +191,7 @@ TEST(FastVector_Tests, Clear_Trivial_Empty)
 
   vector.Clear();
 
-  EXPECT_EQ(vector.Size(), 0);
+  EXPECT_EQ(vector.Size(), 0u);
   EXPECT_TRUE(vector.Empty());
 }
 
@@ -204,7 +204,7 @@ TEST(FastVector_Tests, Clear_NonTrivial_Empty)
 
   vector.Clear();
 
-  EXPECT_EQ(vector.Size(), 0);
+  EXPECT_EQ(vector.Size(), 0u);
   EXPECT_TRUE(vector.Empty());
 }
 
@@ -214,7 +214,7 @@ TEST(FastVector_Tests, Reserve_Trivial_Empty_CorrectCapacity)
 
   vector.Reserve(10);
 
-  EXPECT_EQ(vector.Capacity(), 10);
+  EXPECT_EQ(vector.Capacity(), 10u);
 }
 
 TEST(FastVector_Tests, Reserve_NonTrivial_Empty_CorrectCapacity)
@@ -223,7 +223,7 @@ TEST(FastVector_Tests, Reserve_NonTrivial_Empty_CorrectCapacity)
 
   vector.Reserve(10);
 
-  EXPECT_EQ(vector.Capacity(), 10);
+  EXPECT_EQ(vector.Capacity(), 10u);
 }
 
 TEST(FastVector_Tests, Reserve_Trivial_Increase_CorrectCapacity)
@@ -232,11 +232,11 @@ TEST(FastVector_Tests, Reserve_Trivial_Increase_CorrectCapacity)
 
   vector.Reserve(5);
 
-  EXPECT_EQ(vector.Capacity(), 5);
+  EXPECT_EQ(vector.Capacity(), 5u);
 
   vector.Reserve(10);
 
-  EXPECT_EQ(vector.Capacity(), 10);
+  EXPECT_EQ(vector.Capacity(), 10u);
 }
 
 TEST(FastVector_Tests, Reserve_NonTrivial_Increase_CorrectCapacity)
@@ -245,11 +245,11 @@ TEST(FastVector_Tests, Reserve_NonTrivial_Increase_CorrectCapacity)
 
   vector.Reserve(5);
 
-  EXPECT_EQ(vector.Capacity(), 5);
+  EXPECT_EQ(vector.Capacity(), 5u);
 
   vector.Reserve(10);
 
-  EXPECT_EQ(vector.Capacity(), 10);
+  EXPECT_EQ(vector.Capacity(), 10u);
 }
 
 TEST(FastVector_Tests, Reserve_Trivial_Decrease_DoNothing)
@@ -258,11 +258,11 @@ TEST(FastVector_Tests, Reserve_Trivial_Decrease_DoNothing)
 
   vector.Reserve(10);
 
-  EXPECT_EQ(vector.Capacity(), 10);
+  EXPECT_EQ(vector.Capacity(), 10u);
 
   vector.Reserve(5);
 
-  EXPECT_EQ(vector.Capacity(), 10);
+  EXPECT_EQ(vector.Capacity(), 10u);
 }
 
 TEST(FastVector_Tests, Reserve_NonTrivial_Decrease_DoNothing)
@@ -271,11 +271,11 @@ TEST(FastVector_Tests, Reserve_NonTrivial_Decrease_DoNothing)
 
   vector.Reserve(10);
 
-  EXPECT_EQ(vector.Capacity(), 10);
+  EXPECT_EQ(vector.Capacity(), 10u);
 
   vector.Reserve(5);
 
-  EXPECT_EQ(vector.Capacity(), 10);
+  EXPECT_EQ(vector.Capacity(), 10u);
 }
 
 TEST(FastVector_Tests, PopBack_NonTrivial_Single_SizeDecrease)
@@ -285,7 +285,7 @@ TEST(FastVector_Tests, PopBack_NonTrivial_Single_SizeDecrease)
   vector.PushBack("0");
   vector.PopBack();
 
-  EXPECT_EQ(vector.Size(), 0);
+  EXPECT_EQ(vector.Size(), 0u);
   EXPECT_TRUE(vector.Empty());
 }
 
@@ -298,7 +298,7 @@ TEST(FastVector_Tests, PopBack_Trivial_Double_SizeDecrease)
   vector.PopBack();
   vector.PopBack();
 
-  EXPECT_EQ(vector.Size(), 0);
+  EXPECT_EQ(vector.Size(), 0u);
   EXPECT_TRUE(vector.Empty());
 }
 
@@ -311,7 +311,7 @@ TEST(FastVector_Tests, PopBack_NonTrivial_Double_SizeDecrease)
   vector.PopBack();
   vector.PopBack();
 
-  EXPECT_EQ(vector.Size(), 0);
+  EXPECT_EQ(vector.Size(), 0u);
   EXPECT_TRUE(vector.Empty());
 }
 
@@ -323,7 +323,7 @@ TEST(FastVector_Tests, PopBack_Trivial_DoublePushSinglePop_SizeDecrease)
   vector.PushBack(0);
   vector.PopBack();
 
-  EXPECT_EQ(vector.Size(), 1);
+  EXPECT_EQ(vector.Size(), 1u);
   EXPECT_FALSE(vector.Empty());
 }
 
@@ -335,7 +335,7 @@ TEST(FastVector_Tests, PopBack_NonTrivial_DoublePushSinglePop_SizeDecrease)
   vector.PushBack("0");
   vector.PopBack();
 
-  EXPECT_EQ(vector.Size(), 1);
+  EXPECT_EQ(vector.Size(), 1u);
   EXPECT_FALSE(vector.Empty());
 }
 
@@ -369,7 +369,7 @@ TEST(FastVector_Tests, PopBack_Trivial_PushPopPush_CorrectValues)
   vector.PopBack();
   vector.PushBack(2);
 
-  EXPECT_EQ(vector.Size(), 1);
+  EXPECT_EQ(vector.Size(), 1u);
   EXPECT_EQ(vector[0], 2);
 }
 
@@ -381,7 +381,7 @@ TEST(FastVector_Tests, PopBack_NonTrivial_PushPopPush_CorrectSize)
   vector.PopBack();
   vector.PushBack("2");
 
-  EXPECT_EQ(vector.Size(), 1);
+  EXPECT_EQ(vector.Size(), 1u);
   EXPECT_EQ(vector[0], std::string { "2" });
 }
 
@@ -394,7 +394,7 @@ TEST(FastVector_Tests, PopBack_Trivial_PushPushPopPush_CorrectValues)
   vector.PopBack();
   vector.PushBack(3);
 
-  EXPECT_EQ(vector.Size(), 2);
+  EXPECT_EQ(vector.Size(), 2u);
   EXPECT_EQ(vector[1], 3);
 }
 
@@ -407,7 +407,7 @@ TEST(FastVector_Tests, PopBack_NonTrivial_PushPushPopPush_CorrectValues)
   vector.PopBack();
   vector.PushBack("3");
 
-  EXPECT_EQ(vector.Size(), 2);
+  EXPECT_EQ(vector.Size(), 2u);
   EXPECT_EQ(vector[1], std::string { "3" });
 }
 
@@ -418,7 +418,7 @@ TEST(FastVector_Tests, EraseAt_Trivial_Single_CorrectValues)
   vector.PushBack(1);
   vector.EraseAt(0);
 
-  EXPECT_EQ(vector.Size(), 0);
+  EXPECT_EQ(vector.Size(), 0u);
   EXPECT_TRUE(vector.Empty());
 }
 
@@ -429,7 +429,7 @@ TEST(FastVector_Tests, EraseAt_NonTrivial_Single_CorrectValues)
   vector.PushBack("1");
   vector.EraseAt(0);
 
-  EXPECT_EQ(vector.Size(), 0);
+  EXPECT_EQ(vector.Size(), 0u);
   EXPECT_TRUE(vector.Empty());
 }
 
@@ -441,13 +441,13 @@ TEST(FastVector_Tests, EraseAt_Trivial_Double_CorrectValues)
   vector.PushBack(2);
   vector.EraseAt(1);
 
-  EXPECT_EQ(vector.Size(), 1);
+  EXPECT_EQ(vector.Size(), 1u);
   EXPECT_FALSE(vector.Empty());
   EXPECT_EQ(vector[0], 1);
 
   vector.EraseAt(0);
 
-  EXPECT_EQ(vector.Size(), 0);
+  EXPECT_EQ(vector.Size(), 0u);
   EXPECT_TRUE(vector.Empty());
 }
 
@@ -459,13 +459,13 @@ TEST(FastVector_Tests, EraseAt_NonTrivial_Double_CorrectValues)
   vector.PushBack("2");
   vector.EraseAt(1);
 
-  EXPECT_EQ(vector.Size(), 1);
+  EXPECT_EQ(vector.Size(), 1u);
   EXPECT_FALSE(vector.Empty());
   EXPECT_EQ(vector[0], "1");
 
   vector.EraseAt(0);
 
-  EXPECT_EQ(vector.Size(), 0);
+  EXPECT_EQ(vector.Size(), 0u);
   EXPECT_TRUE(vector.Empty());
 }
 
@@ -477,7 +477,7 @@ TEST(FastVector_Tests, EraseAt_Trivial_PushPushErase_CorrectValues)
   vector.PushBack(2);
   vector.EraseAt(0);
 
-  EXPECT_EQ(vector.Size(), 1);
+  EXPECT_EQ(vector.Size(), 1u);
   EXPECT_FALSE(vector.Empty());
   EXPECT_EQ(vector[0], 2);
 }
@@ -490,7 +490,7 @@ TEST(FastVector_Tests, EraseAt_NonTrivial_PushPushErase_CorrectValues)
   vector.PushBack("2");
   vector.EraseAt(0);
 
-  EXPECT_EQ(vector.Size(), 1);
+  EXPECT_EQ(vector.Size(), 1u);
   EXPECT_FALSE(vector.Empty());
   EXPECT_EQ(vector[0], "2");
 }
@@ -504,7 +504,7 @@ TEST(FastVector_Tests, Erase_Trivial_PushPushErase_CorrectValues)
 
   vector.Erase(std::find(vector.begin(), vector.end(), 1));
 
-  EXPECT_EQ(vector.Size(), 1);
+  EXPECT_EQ(vector.Size(), 1u);
   EXPECT_FALSE(vector.Empty());
   EXPECT_EQ(vector[0], 2);
 }
@@ -518,7 +518,7 @@ TEST(FastVector_Tests, Erase_NonTrivial_PushPushErase_CorrectValues)
 
   vector.Erase(std::find(vector.begin(), vector.end(), std::string { "1" }));
 
-  EXPECT_EQ(vector.Size(), 1);
+  EXPECT_EQ(vector.Size(), 1u);
   EXPECT_FALSE(vector.Empty());
   EXPECT_EQ(vector[0], std::string { "2" });
 }
@@ -629,14 +629,14 @@ TEST(FastVector_Tests, Resize_Trivial_Decrease_CorrectValues)
 
   vector.Resize(2);
 
-  EXPECT_EQ(vector.Size(), 2);
+  EXPECT_EQ(vector.Size(), 2u);
   EXPECT_EQ(vector[0], 1);
   EXPECT_EQ(vector[1], 2);
 
   vector.PushBack(4);
   vector.Resize(1);
 
-  EXPECT_EQ(vector.Size(), 1);
+  EXPECT_EQ(vector.Size(), 1u);
   EXPECT_EQ(vector[0], 1);
 }
 
@@ -650,14 +650,14 @@ TEST(FastVector_Tests, Resize_NonTrivial_Decrease_CorrectValues)
 
   vector.Resize(2);
 
-  EXPECT_EQ(vector.Size(), 2);
+  EXPECT_EQ(vector.Size(), 2u);
   EXPECT_EQ(vector[0], std::string { "1" });
   EXPECT_EQ(vector[1], std::string { "2" });
 
   vector.PushBack("4");
   vector.Resize(1);
 
-  EXPECT_EQ(vector.Size(), 1);
+  EXPECT_EQ(vector.Size(), 1u);
   EXPECT_EQ(vector[0], std::string { "1" });
 }
 
@@ -791,7 +791,7 @@ TEST(FastVector_Tests, Back_Trivial_CorrectValue)
     vector.PushBack(static_cast<double>(i));
   }
 
-  EXPECT_EQ(vector.back(), cAmount - 1);
+  EXPECT_EQ(vector.back(), static_cast<double>(cAmount - 1UL));
 }
 
 TEST(FastVector_Tests, MoveConstructor_Trivial_CorrectValues)
@@ -803,9 +803,9 @@ TEST(FastVector_Tests, MoveConstructor_Trivial_CorrectValues)
 
   FastVector<double> copy { std::move(vector) };
 
-  EXPECT_EQ(copy.Size(), 2);
-  EXPECT_EQ(vector.Size(), 0);
-  EXPECT_EQ(vector.Capacity(), 0);
+  EXPECT_EQ(copy.Size(), 2u);
+  EXPECT_EQ(vector.Size(), 0u);
+  EXPECT_EQ(vector.Capacity(), 0u);
 
   EXPECT_EQ(copy[0], 1);
   EXPECT_EQ(copy[1], 2);
@@ -820,9 +820,9 @@ TEST(FastVector_Tests, MoveConstructor_NonTrivial_CorrectValues)
 
   FastVector<std::string> copy { std::move(vector) };
 
-  EXPECT_EQ(copy.Size(), 2);
-  EXPECT_EQ(vector.Size(), 0);
-  EXPECT_EQ(vector.Capacity(), 0);
+  EXPECT_EQ(copy.Size(), 2u);
+  EXPECT_EQ(vector.Size(), 0u);
+  EXPECT_EQ(vector.Capacity(), 0u);
 
   EXPECT_EQ(copy[0], std::string { "1" });
   EXPECT_EQ(copy[1], std::string { "2" });
@@ -837,9 +837,9 @@ TEST(FastVector_Tests, MoveAssignment_Trivial_CorrectValues)
 
   FastVector<double> moved = std::move(vector);
 
-  EXPECT_EQ(moved.Size(), 2);
-  EXPECT_EQ(vector.Size(), 0);
-  EXPECT_EQ(vector.Capacity(), 0);
+  EXPECT_EQ(moved.Size(), 2u);
+  EXPECT_EQ(vector.Size(), 0u);
+  EXPECT_EQ(vector.Capacity(), 0u);
 
   EXPECT_EQ(moved[0], 1);
   EXPECT_EQ(moved[1], 2);
@@ -854,9 +854,9 @@ TEST(FastVector_Tests, MoveAssignment_NonTrivial_CorrectValues)
 
   FastVector<std::string> moved = std::move(vector);
 
-  EXPECT_EQ(moved.Size(), 2);
-  EXPECT_EQ(vector.Size(), 0);
-  EXPECT_EQ(vector.Capacity(), 0);
+  EXPECT_EQ(moved.Size(), 2u);
+  EXPECT_EQ(vector.Size(), 0u);
+  EXPECT_EQ(vector.Capacity(), 0u);
 
   EXPECT_EQ(moved[0], std::string { "1" });
   EXPECT_EQ(moved[1], std::string { "2" });
@@ -871,7 +871,7 @@ TEST(FastVector_Tests, MoveAssignment_SelfMove_DoNothing)
 
   vector = std::move(vector);
 
-  EXPECT_EQ(vector.Size(), 2);
+  EXPECT_EQ(vector.Size(), 2u);
   EXPECT_EQ(vector[0], 1);
   EXPECT_EQ(vector[1], 2);
 }
@@ -885,9 +885,9 @@ TEST(FastVector_Tests, CopyConstructor_Trivial_CorrectValues)
 
   FastVector<double> copy { vector };
 
-  EXPECT_EQ(copy.Size(), 2);
-  EXPECT_EQ(copy.Capacity(), 2);
-  EXPECT_EQ(vector.Size(), 2);
+  EXPECT_EQ(copy.Size(), 2u);
+  EXPECT_EQ(copy.Capacity(), 2u);
+  EXPECT_EQ(vector.Size(), 2u);
 
   EXPECT_EQ(vector[0], 1);
   EXPECT_EQ(vector[1], 2);
@@ -904,9 +904,9 @@ TEST(FastVector_Tests, CopyConstructor_NonTrivial_CorrectValues)
 
   FastVector<std::string> copy { vector };
 
-  EXPECT_EQ(copy.Size(), 2);
-  EXPECT_EQ(copy.Capacity(), 2);
-  EXPECT_EQ(vector.Size(), 2);
+  EXPECT_EQ(copy.Size(), 2u);
+  EXPECT_EQ(copy.Capacity(), 2u);
+  EXPECT_EQ(vector.Size(), 2u);
 
   EXPECT_EQ(vector[0], std::string { "1" });
   EXPECT_EQ(vector[1], std::string { "2" });
@@ -923,9 +923,9 @@ TEST(FastVector_Tests, CopyAssignment_Trivial_CorrectValues)
 
   FastVector<double> copy = vector;
 
-  EXPECT_EQ(copy.Size(), 2);
-  EXPECT_EQ(copy.Capacity(), 2);
-  EXPECT_EQ(vector.Size(), 2);
+  EXPECT_EQ(copy.Size(), 2u);
+  EXPECT_EQ(copy.Capacity(), 2u);
+  EXPECT_EQ(vector.Size(), 2u);
 
   EXPECT_EQ(vector[0], 1);
   EXPECT_EQ(vector[1], 2);
@@ -942,9 +942,9 @@ TEST(FastVector_Tests, CopyAssignment_NonTrivial_CorrectValues)
 
   FastVector<std::string> copy = vector;
 
-  EXPECT_EQ(copy.Size(), 2);
-  EXPECT_EQ(copy.Capacity(), 2);
-  EXPECT_EQ(vector.Size(), 2);
+  EXPECT_EQ(copy.Size(), 2u);
+  EXPECT_EQ(copy.Capacity(), 2u);
+  EXPECT_EQ(vector.Size(), 2u);
 
   EXPECT_EQ(vector[0], std::string { "1" });
   EXPECT_EQ(vector[1], std::string { "2" });
@@ -961,7 +961,7 @@ TEST(FastVector_Tests, CopyAssignment_SelfAssignement_DoNothing)
 
   vector = vector;
 
-  EXPECT_EQ(vector.Size(), 2);
+  EXPECT_EQ(vector.Size(), 2u);
   EXPECT_EQ(vector[0], 1);
   EXPECT_EQ(vector[1], 2);
 }
